requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pandas
python-dotenv
pydantic-settings
python-multipart

